now full working support for windows
windows is a lot more whinny
it wants some data variables to be null before they are filled in with info

also apparently free() will segfault/sigsev when  you point it towards the middle of an alloc

+++ include/soc.h
@@ -5,7 +5,7 @@
 
 typedef struct {
     CWeb3Config config;
-    int socket;
+    uint64_t socket;
 } CWeb3Socket;
 
 CWeb3Socket newCWeb3Socket(CWeb3Config config);
